+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/PrometheusHeadersParser.cs
@@ -165,18 +165,7 @@ internal static class PrometheusHeadersParser
         var trimmed = TrimQuotes(value);
         var escaping = trimmed.ToString();
 
-        if (PrometheusProtocol.SupportedEscapingSchemes.Contains(escaping))
-        {
-            return escaping;
-        }
-
-        // TODO Support other escaping schemes, including at least "allow-utf-8".
-        // For now we treat "allow-utf-8" as if it were "underscores" to avoid fallback
-        // to PrometheusText0.0.4 where it would previously match to OpenMetricsText1.0.0.
-        // See https://github.com/open-telemetry/opentelemetry-dotnet/issues/7246.
-        return string.Equals(escaping, PrometheusProtocol.AllowUtf8Escaping, StringComparison.Ordinal)
-            ? PrometheusProtocol.UnderscoresEscaping
-            : null;
+        return PrometheusProtocol.SupportedEscapingSchemes.Contains(escaping) ? escaping : null;
     }
 
     private static ReadOnlySpan<char> SplitNext(ref ReadOnlySpan<char> span, char character)