+++ test/OpenTelemetry.Exporter.Prometheus.HttpListener.Tests/PrometheusAcceptHeaders.cs
@@ -29,16 +29,6 @@ public static class PrometheusAcceptHeaders
             testCases.Add(accept, "text/plain", false, "0.0.4", null);
         }
 
-        string[] prometheusV1 =
-        [
-            "text/plain;version=1.0.0;escaping=allow-utf-8;q=0.6,*/*;q=0.5",
-        ];
-
-        foreach (var accept in prometheusV1)
-        {
-            testCases.Add(accept, "text/plain", false, "1.0.0", "underscores");
-        }
-
         string[] openMetricsV0 =
         [
             "application/openmetrics-text",
@@ -57,17 +47,13 @@ public static class PrometheusAcceptHeaders
         [
             "application/openmetrics-text; version=1.0.0",
             "application/openmetrics-text; version=\"1.0.0\"",
-            "application/openmetrics-text; version=1.0.0; escaping=allow-utf-8",
             "application/openmetrics-text; version=1.0.0; escaping=underscores",
             "application/openmetrics-text; version=\"1.0.0\"; escaping=\"underscores\"",
             "application/openmetrics-text; version=1.0.0; charset=utf-8",
             "Application/OpenMetrics-Text; version=1.0.0",
             "application/openmetrics-text; version=1.0.0; charset=utf-8; escaping=underscores",
-            "application/openmetrics-text; version=1.0.0; charset=utf-8; escaping=allow-utf-8",
             "text/plain; q=0.3, application/openmetrics-text; version=1.0.0; q=0.9",
             "TEXT/PLAIN; q=0.3, Application/OpenMetrics-Text; version=1.0.0; q=0.9",
-            "application/openmetrics-text;version=1.0.0;escaping=allow-utf-8;q=0.6,*/*;q=0.5",
-            "application/openmetrics-text;version=1.0.0;escaping=allow-utf-8;q=0.6,application/openmetrics-text;version=0.0.1;q=0.5,text/plain;version=1.0.0;escaping=allow-utf-8;q=0.4,text/plain;version=0.0.4;q=0.3,*/*;q=0.2",
             "application/openmetrics-text;version=1.0.0;escaping=underscores;q=0.6,*/*;q=0.5",
             "application/openmetrics-text;version=1.0.0;escaping=underscores;q=0.6,application/openmetrics-text;version=0.0.1;q=0.5,text/plain;version=1.0.0;escaping=allow-utf-8;q=0.4,text/plain;version=0.0.4;q=0.3,*/*;q=0.2",
         ];
@@ -77,7 +63,12 @@ public static class PrometheusAcceptHeaders
             testCases.Add(accept, "application/openmetrics-text", true, "1.0.0", "underscores");
         }
 
-        // The dots and values escaping schemes are negotiated for the only the v1 formats.
+        // The allow-utf-8, dots and values escaping schemes are negotiated for only the v1 formats.
+        testCases.Add("text/plain;version=1.0.0;escaping=allow-utf-8;q=0.6,*/*;q=0.5", "text/plain", false, "1.0.0", "allow-utf-8");
+        testCases.Add("application/openmetrics-text; version=1.0.0; escaping=allow-utf-8", "application/openmetrics-text", true, "1.0.0", "allow-utf-8");
+        testCases.Add("application/openmetrics-text; version=1.0.0; charset=utf-8; escaping=allow-utf-8", "application/openmetrics-text", true, "1.0.0", "allow-utf-8");
+        testCases.Add("application/openmetrics-text;version=1.0.0;escaping=allow-utf-8;q=0.6,*/*;q=0.5", "application/openmetrics-text", true, "1.0.0", "allow-utf-8");
+        testCases.Add("application/openmetrics-text;version=1.0.0;escaping=allow-utf-8;q=0.6,application/openmetrics-text;version=0.0.1;q=0.5,text/plain;version=1.0.0;escaping=allow-utf-8;q=0.4,text/plain;version=0.0.4;q=0.3,*/*;q=0.2", "application/openmetrics-text", true, "1.0.0", "allow-utf-8");
         testCases.Add("application/openmetrics-text; version=1.0.0; escaping=dots", "application/openmetrics-text", true, "1.0.0", "dots");
         testCases.Add("application/openmetrics-text; version=1.0.0; charset=utf-8; escaping=dots", "application/openmetrics-text", true, "1.0.0", "dots");
         testCases.Add("application/openmetrics-text; version=1.0.0; escaping=values", "application/openmetrics-text", true, "1.0.0", "values");